+++ packages/components/src/ProForm/formdom.tsx
@@ -34,7 +34,6 @@ function FormDom({
   useEffect(() => {
     formInstanceRef.current = baseRef;
   }, [baseRef]);
-
   return (
     <Form
       ref={baseRef}